adds permission evaluator interface for QAI

+++ module/src/main/java/net/savantly/sprout/franchise/domain/operations/qai/audit/QAAService.java
@@ -10,9 +10,15 @@ import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 import lombok.RequiredArgsConstructor;
+import lombok.var;
 import lombok.extern.slf4j.Slf4j;
+import net.savantly.sprout.core.security.permissions.Permission;
+import net.savantly.sprout.franchise.domain.operations.qai.QAIPermissionEvaluator;
 import net.savantly.sprout.franchise.domain.operations.qai.score.QAAScoreDto;
 import net.savantly.sprout.franchise.domain.operations.qai.score.QAAScoreService;
 import net.savantly.sprout.franchise.domain.operations.qai.section.submission.QAISectionSubmissionDto;
@@ -25,9 +31,11 @@ public class QAAService {
 	final QAISubmissionService qaiSubmissionService;
 	final QAASubmissionRepository repository;
 	final QAAScoreService scoreService;
+	final QAIPermissionEvaluator permissionEvaluator;
 	
 	public Page<QAASummaryDto> getPage(Pageable pageable) {
-		return this.repository.findAll(pageable).map(q -> convertToSummary(q));
+		var pageOfSummaries = this.repository.findAll(pageable).map(q -> convertToSummary(q));
+		return permissionEvaluator.filterByPermission(SecurityContextHolder.getContext().getAuthentication(), pageOfSummaries);
 	}
 
 	private QAASummaryDto convertToSummary(QAASubmission q) {
@@ -45,10 +53,19 @@ public class QAAService {
 
 	public QAADto getOneById(String id) {
 		log.info("finding QAAScore by id: {}", id);
-		return convert(this.repository.findById(id).orElseThrow());
+		var dto = convert(this.repository.findById(id).orElseThrow());
+		if (permissionEvaluator.hasPermission(getAuth(), dto, Permission.READ)) {
+			return dto;
+		} else {
+			throw new AccessDeniedException("not authorized to READ the record with ID: " + dto.getId());
+		}
 	}
 	
+
 	public QAADto save(QAADto dto) {
+		if (!permissionEvaluator.hasPermission(getAuth(), dto, Permission.UPDATE)) {
+			throw new AccessDeniedException("not authorized to UPDATE the record with ID: " + dto.getId());
+		}
 		Set<String> sectionSubmissionIds = saveSectionSubmissions(dto);
 		QAASubmission entity = convert(dto);
 		entity.setSectionSubmissionIds(sectionSubmissionIds);
@@ -60,6 +77,9 @@ public class QAAService {
 	}
 	
 	public void delete(String id) {
+		if (!permissionEvaluator.hasDeletePermission(getAuth(), id)) {
+			throw new AccessDeniedException("not authorized to DELETE the record with ID: " + id);
+		}
 		this.repository.deleteById(id);
 	}
 
@@ -124,6 +144,10 @@ public class QAAService {
 		QAASubmission entity = new QAASubmission();
 		entity.setId(UUID.randomUUID().toString());
 		return entity;
+	}
+
+	private Authentication getAuth() {
+		return SecurityContextHolder.getContext().getAuthentication();
 	}
 
 }
